192 Profile Support WIP (#211)
* Functionality for the Reset HUD button

Changed the way configurations are read and saved to implement the reset-to-default functionality

* Functionality for import/export

New GUI tab for importing/export configurations. For export: PluginConfiguration is serialized to JSON, compressed by DeflateStream, then encoded in base64. Import is the reverse.
Not all parameters are properly written/read, so more work is needed.

* Removed extraneous assembly reference

* Import and export now read fields, not properties

Also added a paste-from-clipboard button for importing.