Fix PrimaryResourceBar display configuration
Any modifications wherever to display or not the Primary Resource Bar in
the job configurations was not being processed until the player changed
jobs.

This is due to the fact that there's 2 conditions wherever that
PlayerResourceBar should be displayed:
- In the Misc configuration, checked everytime the PrimaryResourceHud
  Draw call si done
- In the Job configuration

The Job configuration is being used for the PrimaryReshourceHud when the
UpdateJob method is called in HudManager.cs.
Except, that configuration is then only processed when the JobHud isn't
created or when the player changes job.

This modificaiton forces to read again the Job configuration when the
player doesn't change jobs, making both configuration work as intended.