Unitframes with new bars (#570)
Rewrite of UnitframeHud to use the new barhud for unitframes.

note: I completely removed the tank stance indicator because the old one was problematic (clipping issues and doesn't work with unitframes that can now be transparent). Still need a new design for this, but for now tracking with the buffs feature should work well enough.